+++ package.json
@@ -8,9 +8,11 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.5.0",
     "bootstrap": "^5.3.1",
+    "http-proxy-middleware": "^2.0.6",
     "next": "^13.4.19",
     "react": "^18.2.0",
     "react-bootstrap": "^2.8.0",
+    "react-cookie": "^6.1.1",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
