+++ src/component/Header.js
@@ -2,8 +2,8 @@ import { React, useState } from 'react';
 import {Navbar, Nav, Container, NavDropdown} from 'react-bootstrap'
 import '../bootstrap.css';
 import { Link } from 'react-router-dom';
-import Login from '../pages/Login';
-import Regist from '../pages/Regist';
+import RegistModal from './Modal/RegistModal';
+import LoginModal from './Modal/LoginModal';
 
 
 function Header() {
@@ -26,9 +26,9 @@ function Header() {
 					<div className="mainButton"> 웹툰 그림체 학습</div>
             <Nav className="ml-auto">
               <button onClick={openLoginModal} className="header-button">LOGIN</button>
-              {isLoginOpen && (<Login isOpen={isLoginOpen} closeModal={closeLoginModal}/>)}
+              {isLoginOpen && (<LoginModal isOpen={isLoginOpen} closeModal={closeLoginModal}/>)}
               <button onClick={openRegistModal} className="header-button">SIGN UP</button>
-              {isRegistOpen && (<Regist isOpen={isRegistOpen} closeModal={closeRegistModal}/>)}
+              {isRegistOpen && (<RegistModal isOpen={isRegistOpen} closeModal={closeRegistModal}/>)}
                 <form className="d-flex">
                     <input class="form-control me-sm-2 search" type="search" placeholder="Search"/>
                     <button type="submit" className="">Search</button>
