+++ src/component/Modal/RegistModal.js
@@ -1,0 +1,111 @@
+import { React, useEffect, useState } from 'react'
+import '../../styles/pages.css';
+
+
+const User = {
+  email: 'test@example.com',
+  pw: 'test2323@@@'
+}
+
+function Regist({isOpen, closeModal}) {
+
+  const [email, setEmail] = useState('');
+  const [pw, setPw] = useState("");
+
+  const [emailValid, setEmailValid] = useState(false);
+  const [pwValid, setPwValid] = useState(false);
+  const [notAllow, setNotAllow] = useState(true);
+
+  const handleEmail = (e) => {
+    setEmail(e.target.value);
+    const regex =
+      /^(([^<>()\[\].,;:\s@"]+(\.[^<>()\[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/i;
+      // /^[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_\.]?[0-9a-zA-Z])*\.[a-zA-Z]{2,3}$/i;
+    if(regex.test(email)) {
+      setEmailValid(true);
+    } else {
+      setEmailValid(false);
+    }
+  }
+
+  const handlePw= (e) => {
+    setPw(e.target.value);
+    const rePass =  /^[a-zA-Z\\d`~!@#$%^&*()-_=+]{8,24}$/;
+    if(rePass.test(pw)) {
+      setPwValid(true);
+    } else {
+      setPwValid(false);
+    }
+  }
+
+  const onClickConfirmButton = () => {
+    if(email === User.email && pw === User.pw) {
+      alert('로그인에 성공했습니다.');
+    } else {
+      alert('등록되지 않은 회원입니다.')
+    }
+  }
+
+  useEffect(() => {
+    if(emailValid && pwValid) {
+      setNotAllow(false);
+      return;
+    }
+      setNotAllow(true);
+  }, [emailValid, pwValid])
+
+  return (
+      <div className="loginPage" style={{display:isOpen?"block":"none",}}>
+        <div className="loginModal">
+          <div className="titleWrap">
+            SIGN UP
+          </div>
+          <div className="contentWrap">
+            <div className="inputTitle">이메일 주소</div>
+            <div className="inputWrap">
+              <input 
+                type='text'
+                className='input' 
+                placeholder='test@gmail.com' 
+                value={email} 
+                onChange={handleEmail}/>
+            </div>
+            <div className="errorMessageWrap">
+              {
+                !emailValid && email.length > 0 && (
+                  <div>올바른 이메일을 입력해주세요.</div>
+                )
+              }
+            </div>
+            <div style={{ marginTop: "26px" }}className="inputTitle">비밀번호</div>
+            <div className="inputWrap">
+              <input
+                type='password' 
+                className='input' 
+                placeholder='영문,숫자,특수문자 포함 8자 이상'
+                value={pw}
+                onChange={handlePw}/>
+            </div>  
+            <div className="errorMessageWrap">
+              {
+                !pwValid && pw.length > 0 && (
+                  <div>영문,숫자,특수문자 포함 8자 이상을 입력해주세요.</div>
+                )
+              }
+            </div>
+          </div>
+
+          <div className="buttonBox">
+            <button onClick={onClickConfirmButton} disabled={notAllow} className='loginButton'> 
+              확인
+            </button>
+            <button onClick={closeModal} className='buttonBox'> 
+              닫기
+            </button>
+          </div>
+        </div>
+      </div>
+  )
+}
+
+export default Regist
