+++ src/setProxy.js
@@ -1,0 +1,10 @@
+const { createProxyMiddleware } = require("http-proxy-middleware");
+
+module.exports = function (app) {
+  app.use(
+    createProxyMiddleware("/api/v1", {
+      target: "http://localhost:8080",
+      changeOrigin: true,
+    })
+  );
+};
