+++ src/styles/NovelCharacter.css
